requirements.txt
streamlit
PyPDF2
python-docx
python-pptx
pandas
transformers
datasets
scikit-learn
torch
sentence-transformers